Exercises 4.18 to 4.23

+++ part4/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -10,6 +9,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+
   if (!body.title) {
     return response.status(400).json({
       error: 'title is missing!',
@@ -21,11 +21,7 @@ blogsRouter.post('/', async (request, response) => {
     })
   }
 
-  const user = await User.findOne({})
-
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
-  }
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -44,8 +40,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request,response) => {
+  const user = request.user
   const id = request.params.id
+  const blogToDelete = await Blog.findById(id)
+
+  if (blogToDelete.user.toString() !== user.id.toString()){
+    return response.status(401).json({ error: 'not authorized to delete this blog' })
+  }
+
   await Blog.findByIdAndDelete(id)
+
   response.status(204).end()
 })
